feat: a confirmation dialog was created when deleting a course

+++ src/app/courses/containers/courses/courses.component.ts
@@ -12,6 +12,7 @@ import { Course } from '../../model/course';
 import { CoursesService } from '../../services/courses.service';
 import { CoursesListComponent } from "../../components/courses-list/courses-list.component";
 import { MatSnackBar } from '@angular/material/snack-bar';
+import { ConfirmationDialogComponent } from '../../../shared/components/confirmation-dialog/confirmation-dialog.component';
 
 @Component({
     selector: 'app-courses',
@@ -64,17 +65,29 @@ export class CoursesComponent {
   }
 
   onRemove(course: Course){
-    this.coursesService.remove(course._id)
-      .subscribe({
-        next: () => {
-          this.refresh()
-          this.snackBar.open("Curso removido com sucesso!", 'X', {
-            duration: 5000,
-            verticalPosition: "top",
-            horizontalPosition: "center"
-          })
-        },
-        error: () => this.onError("Erro ao tentar remover curso!")
-      })
+    const dialogRef = this.dialog.open(ConfirmationDialogComponent, {
+      data: "Tem certeza que deseja remover esse curso?",
+    });
+
+    dialogRef.afterClosed().subscribe((result: boolean) => {
+      if(result){
+        this.coursesService.remove(course._id)
+          .subscribe({
+            next: () => {
+            this.refresh()
+            this.snackBar.open("Curso removido com sucesso!", 'X', {
+              duration: 5000,
+              verticalPosition: "top",
+              horizontalPosition: "center"
+            })
+          },
+          error: () => this.onError("Erro ao tentar remover curso!")
+        })
+      }
+    });
+
+
+
   }
+
 }
